refactor: Prepare lexer for chumsky 0.10 update
Add feature flag chumsky-10 that when active, makes the lexer use a different implementation.
For now, that implementation is a stub that throws an error, but this sets up the structure to allow
incrementally building the chumsky 0.10 lexer while keeping the 0.9 one working.
