+++ prqlc/prqlc-parser/Cargo.toml
@@ -12,6 +12,10 @@ version.workspace = true
 bench = false
 doctest = false
 
+[features]
+# Enable chumsky-10 to use Chumsky 0.10.0
+chumsky-10 = []
+
 [dependencies]
 enum-as-inner = {workspace = true}
 itertools = {workspace = true}
@@ -26,7 +30,7 @@ strum = {version = "0.27.0", features = ["std", "derive"]}
 # see it when compiling on macOS), so we only include features when running
 # outside wasm.
 [target.'cfg(not(target_family="wasm"))'.dependencies]
-chumsky = "0.9.2"
+chumsky = { version = "0.9.2" }
 # Not direct dependencies, but pinning because of bugs in previous versions
 stacker = "0.1.18"
 [target.'cfg(target_family="wasm")'.dependencies]
