+++ prqlc/prqlc-parser/src/error.rs
@@ -1,6 +1,5 @@
 use std::fmt::Debug;
 
-use chumsky::error::Cheap;
 use serde::Serialize;
 
 use super::parser::perror::PError;
@@ -19,15 +18,30 @@ pub struct Error {
     // pub source: ErrorSource
 }
 
+#[cfg(not(feature = "chumsky-10"))]
 #[derive(Clone, Debug, Default)]
 pub enum ErrorSource {
-    Lexer(Cheap<char>),
+    Lexer(chumsky::error::Cheap<char>),
     Parser(PError),
     #[default]
     Unknown,
     NameResolver,
     TypeResolver,
     SQL,
+    Internal { message: String },
+}
+
+#[cfg(feature = "chumsky-10")]
+#[derive(Clone, Debug, Default)]
+pub enum ErrorSource {
+    Lexer(String), // We'll store the error message as a string since we can't easily store the error type
+    Parser(PError),
+    #[default]
+    Unknown,
+    NameResolver,
+    TypeResolver,
+    SQL,
+    Internal { message: String },
 }
 
 /// Multiple prqlc errors. Used internally, exposed as prqlc::ErrorMessages.
@@ -64,6 +78,9 @@ pub enum Reason {
     Bug {
         issue: Option<i32>,
         details: Option<String>,
+    },
+    Internal {
+        message: String,
     },
 }
 
@@ -127,6 +144,9 @@ impl std::fmt::Display for Reason {
                     )?;
                 }
                 Ok(())
+            }
+            Reason::Internal { message } => {
+                write!(f, "internal error: {message}")
             }
         }
     }
